requirements.txt
asgiref==3.4.1
astroid==2.9.0
Django==4.0
django-crispy-forms==1.13.0
django-filter==21.1
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.0
psycopg2==2.9.2
pylint==2.12.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
sqlparse==0.4.2
toml==0.10.2
wrapt==1.13.3
